### 2024-12-02 #7543 @editorialbot
[PRE REVIEW]: ConleyDynamics.jl: A Julia package for multivector dynamics on Lefschetz complexes
**Submitting author:** @almost6heads (<a href="http://orcid.org/0000-0003-3294-0366">Thomas Wanner</a>)
**Repository:** https://github.com/almost6heads/ConleyDynamics.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
