### 2024-12-02 #7543 @danielskatz
👋 @almost6heads - Thanks for your submission.  It appears a bit long, about 1900 words, while JOSS typically expects 250-1000. Is there anything you can move to the repo or documentation and just point to or summarize in the paper?

In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
