### 2024-12-02 #7543 @danielskatz
Also, I see in the [README](https://github.com/almost6heads/ConleyDynamics.jl?tab=readme-ov-file#contributing)
> Currently, this repository does not allow any outside contributors. But if you are interested in getting involved, let me know!

This will need to be changed for JOSS to something more welcoming, as described in https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines
