### 2024-12-02 #7543 @almost6heads
@danielskatz I have added guideline for contributing to the project.
Also, possible reviewers might be: pitmonticone, DanielVandH, avik-pal, LuisScoccola.

I will work on the missing DOIs, and will try to shorten the paper as soon as I can. Thanks!
