### 2024-12-02 #7543 @almost6heads
@danielskatz I am almost ready to move forward, but do have one question: I located three of the missing DOIs, but there is still one labeled as missing. The problem is that this book does not have a DOI, but there is an unrelated paper with the same title that does. How can I deal with that?
Other than that, the paper has been shortened.
