### 2024-12-02 #7543 @danielskatz
And note that the author suggested some possible reviewers in https://github.com/openjournals/joss-reviews/issues/7543#issuecomment-2512161224
