### 2025-01-17 #7543 @drvinceknight
Hi @pitmonticone, @DanielVandH, @avik-pal, @LuisScoccola I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “ ConleyDynamics.jl: A Julia package for multivector dynamics on Lefschetz complexes”. You can find more information at the top of this Github issue.
