### 2025-04-10 #7543 @bonh
Thanks for asking! Up to now I never heard about "combinatorial topological dynamics". However, sounds interesting and relevant, and I would be willing to read up a bit about it. If that would be sufficient for you @dawbarton to review the paper and repository I'd be happy to participate.
