### 2025-04-21 #8085 @editorialbot
[REVIEW]: ConleyDynamics.jl: A Julia package for multivector dynamics on Lefschetz complexes
**Submitting author:** @almost6heads (<a href="http://orcid.org/0000-0003-3294-0366">Thomas Wanner</a>)
**Repository:** https://github.com/almost6heads/ConleyDynamics.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.5-joss
