### 2025-05-16 #8085 @danielskatz
@LuisScoccola & @bonh - Thanks for agreeing to review this submission.

(I'm filling in for @drvinceknight here, to help get things started)

This is the review thread for the paper. All of the communications will happen here from now on.

We aim for reviews to be completed within about 2-4 weeks. Please let @drvinceknight know if either of you require some more time. We can also use editorialbot (our bot) to set automatic reminders if you know you'll be away for a known period of time.

Please feel free to ping @drvinceknight if you have any questions/concerns.
