### 2025-07-05 #8085 @danielskatz
👋 @drvinceknight - it looks like this is ready for you to complete your processing of, and turn it over to me?
