### 2025-07-07 #8085 @drvinceknight
@almost6heads I've made a minor type edit here: https://github.com/almost6heads/ConleyDynamics.jl/pull/5

Otherwise everything looks good (I double checked the DOI suggested by `editorial_bot` and it's not correct -- your citations are all good).

Let me know what you think of my type suggestions, there will be a couple more easy steps and then this will be done.
