### 2025-07-07 #8085 @drvinceknight
Ok great :) Next step: Can you create a new release (if possible) of the software and deposit a new archive (please make sure the authors/title match the title of the paper). Once you have a new DOI can you please let me know what that is.
