+++ v0.3.5-joss/refs.bib
@@ -1,0 +1,290 @@
+
+@article{alexandrov:37a,
+   author = {Alexandrov, P.S.},
+   title = {Diskrete {R}\"aume},
+   journal = {Mathematiceskii Sbornik (N.S.)},
+   volume = {2},
+   year = {1937},
+   pages = {501--518}
+   }
+
+@article{batko:etal:20a,
+   author = {Bogdan Batko and Tomasz Kaczynski and Marian Mrozek and Thomas Wanner},
+   title = {Linking combinatorial and classical dynamics: {C}onley
+            index and {M}orse decompositions},
+   journal = {Foundations of Computational Mathematics},
+   volume = {20},
+   number = {5},
+   year = {2020},
+   pages = {967--1012},
+   doi = {10.1007/s10208-020-09444-1}
+   }
+
+@book{conley:78a,
+   author = {Charles Conley},
+   title = {Isolated Invariant Sets and the {M}orse Index},
+   publisher = {American Mathematical Society},
+   year = 1978,
+   address = {Providence, R.I.}
+   }
+
+@article{dey:etal:24a,
+    AUTHOR = {Dey, Tamal K. and Lipi\'{n}ski, Micha\l and Mrozek, Marian and
+              Slechta, Ryan},
+     TITLE = {Computing connection matrices via persistence-like reductions},
+   JOURNAL = {SIAM Journal on Applied Dynamical Systems},
+    VOLUME = {23},
+      YEAR = {2024},
+    NUMBER = {1},
+     PAGES = {81--97},
+       DOI = {10.1137/23M1562469},
+       }
+
+@article{dlotko:etal:11a,
+   Author = {Pawe\l{} D\l{}otko and Tomasz Kaczynski and
+             Marian Mrozek and Thomas Wanner},
+   title = {Coreduction homology algorithm for regular {C}{W}-complexes},
+   year = 2011,
+   journal = {Discrete \& Computational Geometry},
+   volume = 46,
+   number = 2,
+   pages = {361--388},
+   doi = {10.1007/s00454-010-9303-y}
+   }
+
+@article{dlotko:wanner:18a,
+   author = {Pawe{\l} D{\l}otko and Thomas Wanner},
+   title = {Rigorous cubical approximation and persistent homology of
+            continuous functions},
+   journal = {Computers \& Mathematics with Applications},
+   volume = {75},
+   number = {5},
+   year = {2018},
+   pages = {1648--1666},
+   doi = {10.1016/j.camwa.2017.11.027}
+   }
+
+@book{edelsbrunner:harer:10a,
+   AUTHOR = {Edelsbrunner, Herbert and Harer, John L.},
+   TITLE = {Computational Topology},
+   PUBLISHER = {American Mathematical Society},
+   ADDRESS = {Providence},
+   YEAR = {2010}
+   }
+
+@misc{edelsbrunner:mrozek:23a,
+   author = {Herbert Edelsbrunner and Marian Mrozek},
+   title = {The depth poset of a filtered {L}efschetz complex},
+   year = {2023},
+   eprint = {2311.14364v2},
+   archivePrefix = {arXiv},
+   primaryClass = {math.AT},
+   url = {https://arxiv.org/abs/2311.14364v2}, 
+   }
+
+@article{forman:98a,
+   author = {Forman, Robin},
+   title = {Combinatorial vector fields and dynamical systems},
+   journal = {Mathematische Zeitschrift},
+   volume = {228},
+   year = {1998},
+   number = {4},
+   pages = {629--681},
+   mrnumber = {1644432},
+   doi = {10.1007/PL00004638}
+   }
+
+@article{forman:98b,
+   author = {Forman, Robin},
[…]
