### 2025-07-16 #8085 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
