### 2025-07-22 #8085 @almost6heads
Since my cited book with Marian Mrozek just came out, I added its now available doi to the bib file. Not sure whether the pdf needs to be generated again, or whether this is done later anyway..
