### 2025-07-27 #8085 @danielskatz
@almost6heads - sorry for the delay, I've been on vacation but am now back. I've reviewed the paper, and suggest the changes in https://github.com/almost6heads/ConleyDynamics.jl/pull/6 which are based on how references are directly discussed vs indirect ones.  Please merge this, or let me know what you disagree with.
