### 2025-07-28 #8085 @danielskatz
@almost6heads - please check this proof and let me know if it's ok
