### 2025-07-28 #8085 @almost6heads
@danielskatz I checked the pdf and everything looks good
