### 2025-07-28 #8085 @danielskatz
Sorry, I found one more referencing change I missed, as well as a comma. See https://github.com/almost6heads/ConleyDynamics.jl/pull/7
