### 2025-07-28 #8085 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
