### 2025-07-28 #8085 @danielskatz
Congratulations to @almost6heads (Thomas Wanner) on your JOSS publication!!
